requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
cryptography==42.0.8
datasets==2.20.0
dill==0.3.8
etils==1.9.2
faiss-cpu==1.8.0.post1
filelock==3.15.4
flatbuffers==24.3.25
flax==0.8.5
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.0
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
isodate==0.6.1
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
keras==3.4.1
lamini==2.5.4
lamini-configuration==0.8.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optax==0.2.3
optree==0.12.1
orbax-checkpoint==0.5.20
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sympy==1.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.63
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zstandard==0.22.0
